split routing controllers to smaller modules by function
